feat: 2opt*1

+++ src/main/java/Constraints/Shift20/HardTimeConstraintImpl.java
@@ -10,7 +10,7 @@ public class HardTimeConstraintImpl extends HardTimeConstraint {
     public ConsStatus fulfilled(OperationContext context) {
         Customer cus1 = (Customer) context.mainRoute.getNode(context.operatePos[0]);
         Customer cus2 = (Customer) context.mainRoute.getNode(context.operatePos[0]+1);
-        double timeChg = cus1.duration + cus2.duration;
+        int timeChg = cus1.duration + cus2.duration;
         if (context.sideRoute.getTimeCost() + timeChg > ((Depot) context.sideRoute.start).maxDuration)return ConsStatus.NOT_FULFILLED_BREAK;
         return ConsStatus.FULFILLED;
     }
